requirements/base.txt
# Python Utilities
python-slugify==8.0.4
Pillow==11.1.0
argon2-cffi==23.1.0
pydantic==2.11.1
pydantic-settings==2.8.1
boto3==1.37.24

# Redis
redis==5.3.0
hiredis==3.1.1
psycopg[pool]==3.2.7

# Django and Extensions
django==5.2.0
django-model-utils==5.0.0
django-redis==5.4.0
django-import-export==4.3.7
openpyxl==3.1.5
django-phonenumber-field[phonenumbers]==8.1.0
django-polymorphic==3.1.0
django-unfold==0.57.0
fcm-django==2.2.1
django-solo==2.4.0
rules==3.5
django-imagekit==5.0.0
django-filter==25.1
django-money==3.5.4
django-cors-headers==4.7.0
dj-database-url==2.3.0
django_model_suite==0.2.0
django-simple-history==3.8.0
django-modeltranslation==0.19.14
django-tasks==0.7.0
django-soft-delete==1.0.18
django-guardian==3.0.0

# Django REST Framework and Extensions
djangorestframework==3.15.2 # djangorestframework-gis 1.1 don't accept 3.16
# djangorestframework-gis==1.1
git+https://github.com/openwisp/django-rest-framework-gis.git # noqa

djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
drf-standardized-errors[openapi]==0.14.1
djangorestframework-camel-case==1.4.2
